Change report logic for finding surrounding's of a pracels

+++ packages/Payments/src/DTO/TpayTransactionDTO.php
@@ -6,7 +6,7 @@ final readonly class TpayTransactionDTO
 {
     public function __construct(
         public int $orderId,
-        public int $amount,        // grosze
+        public float $amount,
         public string $email,
         public string $returnUrl,
         public string $notifyUrl,
